requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.49.0
celery==3.1.25
certifi==2018.8.13
cloudinary==1.12.0
Django==1.8
django-celery==3.1.17
djangorestframework==3.1.0
kombu==3.0.37
mock==2.0.0
pbr==4.2.0
Pillow==5.2.0
psycopg2==2.7.4
pytz==2018.5
six==1.11.0
urllib3==1.23